Fix execution stack allocation to avoid unsafe reallocation
Implement the recommendation from the original comment to allocate
a large initial stack instead of dynamically growing it.

Changes:
- Allocate fixed stack of 100,000 elements at initialization
- Remove unsafe reallocation logic that could invalidate pointers
- Add cleanup_cexec() function for proper memory cleanup
- Fix init_cexec() return type from int to long
- Update comments to reflect implemented solution

The dynamic reallocation was unsafe because code execution may store
pointers into the stack. Now we allocate a large stack upfront
(EXEC_STACK_INITIAL_SIZE) to avoid reallocation during execution.

+++ src/viola/viola.c
@@ -47,7 +47,7 @@ extern int init_sys();
 extern int init_obj();
 extern int init_scanutils();
 extern int init_tfed();
-extern int init_cexec();
+extern long init_cexec();
 extern int init_html();
 extern int SGMLInit();
 extern int meth_generic_cli();
